backend/requirements.txt
flask
torch
torchvision
Pillow]
google-generativeai
python-dotenv
flask-cors